Cost fast umgesetzt

+++ TowerDefenseNew/GameObjects/Tower.cs
@@ -12,7 +12,7 @@ namespace TowerDefenseNew.GameObjects
     {
 
 
-        internal Tower(Vector2 center, float attackRadius, int damage, int attackSpeed, int cost, List<Enemy> enemies, List<Bullet> bullets) : base(center, attackRadius)
+        internal Tower(Vector2 center, float attackRadius, int damage, int attackSpeed, double cost, List<Enemy> enemies, List<Bullet> bullets) : base(center, attackRadius)
         {
             this.Center = center;
             this.Enemies = enemies;
@@ -56,7 +56,7 @@ namespace TowerDefenseNew.GameObjects
             }
         }
 
-        private int cost { get; set; }
+        private double cost { get; set; }
         private int attackSpeed { get; set; }
         internal int damage { get; set; }
         private List<Enemy> Enemies { get; set; }
